wip Added ai controller .

+++ src/circular-buffer.h
@@ -16,7 +16,7 @@ bool circularBufferPop(CircularBuffer *circularBuffer, uint8_t *index);
 uint8_t circularBufferSize(CircularBuffer *circularBuffer);
 
 #define circularBufferForEach(circularBuffer, iterator) for (iterator = circularBuffer->begin; iterator != circularBuffer->end; ++iterator)
-#define circularBufferForEachReverse(circularBuffer, iterator) for (iterator = circularBuffer->end - 1; iterator != (uint8_t)(circularBuffer->begin - 1); --iterator)
+#define circularBufferForEachReverse(circularBuffer, iterator) for (iterator = (uint8_t)(circularBuffer->end - 1); iterator != (uint8_t)(circularBuffer->begin - 1); --iterator)
 
 #define circularBufferGetValue(tailValues, index) ((tailValues)[(uint8_t)(index)])
 #define circularBufferGetLastValue(tail, tailValues) circularBufferGetValue(tailValues, tail.end - 1)
